Split XCProject loading to a separate loader
